Refactor ct_parser and ct_imports (#1015)
* Replace inheritance with composition, seperate ct_parser and ct_imports

* Fix sources in ct_import rather than ct_parser

* Fix build

* Fix macos build

* Fix clang build
