+++ future/src/ct/ct_widgets.h
@@ -36,7 +36,6 @@
 #include <map>
 #include "ct_types.h"
 #include "ct_filesystem.h"
-#include "ct_parser.h"
 
 
 class CtMDParser;
